Get permutation matrix from pardiso

+++ SparseLinearSolver/Include/CholeskySolver.h
@@ -12,28 +12,28 @@ namespace Cholesky
 	{
 		int N=0, NZ=0;
 		std::vector<double> Value;
-		std::vector<int> Col, RowIndex;
+		std::vector<int64_t> Col, RowIndex;
 		bool zeroIndexing = true;
 		void PrintFullMatrix()
 		{
-			for (int i = 0; i < RowIndex.size()-1; i++)
+			for (int64_t i = 0; i < int64_t(RowIndex.size()-1); i++)
 			{
-				int startRowIdx = RowIndex[i];
-				int endRowIdx = RowIndex[i + 1] - 1;
+				int64_t startRowIdx = RowIndex[i];
+				int64_t endRowIdx = RowIndex[i + 1] - 1;
 
 				auto startCol = Col[startRowIdx];
-				for (int ii = 0; ii < startCol; ++ii)
+				for (int64_t ii = 0; ii < startCol; ++ii)
 				{
 					std::cout << "      ";
 				}
 				
-				for (int j = startRowIdx; j <= endRowIdx; j++)
+				for (int64_t j = startRowIdx; j <= endRowIdx; j++)
 				{  
 					auto deltaIdx = (Col[j] - startCol)-1;
 					startCol = Col[j];
 					if (deltaIdx > 0)
 					{
-						for (int ii = 0; ii < deltaIdx; ++ii)
+						for (int64_t ii = 0; ii < deltaIdx; ++ii)
 						{
 							std::cout.width(6);
 							std::cout << 0.0;
@@ -44,7 +44,7 @@ namespace Cholesky
 					std::cout << std::setprecision(3)<< Value[j];
 				}
 
-				for (int j = Col[endRowIdx]+1; j < RowIndex.size()-1; ++j)
+				for (int64_t j = Col[endRowIdx]+1; j < int64_t(RowIndex.size()-1); ++j)
 				{
 					std::cout.width(6);
 					std::cout << 0.0;
@@ -61,7 +61,7 @@ namespace Cholesky
 		class Data
 		{
 		public:
-			Data(const std::vector<int>& iColomn, const std::vector<int>& iRowIndex, std::vector<double>& iValues, bool iZeroIndexing);
+			Data(const std::vector<int64_t>& iColomn, const std::vector<int64_t>& iRowIndex, std::vector<double>& iValues, bool iZeroIndexing);
 			
 			void EnableIndexFromZero(bool iFlag) { _enableIndexFromZero = iFlag; };
 		private:
